requirements.txt
python-jsonrpc
pymongo
pika
requests
lxml
sklearn
python-dateutil
newspaper3k